+++ src/tp/logic/IteamUpdateObserver.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface IteamUpdateObserver {
 
-	void onTeamGenerated(List<Employee> team, int combinations, long time);
+	void onTeamGenerated(List<Employee> team, double combinations, long time);
 
 	void onConmparativeGenerated(HashMap<String, Object[]> resultMap);
